refactor(core): split protocol into atoms.rs and loop.rs modules
- Move Atom trait and implementations to new atoms.rs module
- Move AgentLoop2 to new loop.rs module
- protocol.rs now re-exports from atoms and loop modules
- Remove old protocol_agent.rs example
- Create new decomposed_execution.rs example demonstrating direct atom usage
- Add AgentProtocol as type alias for AgentLoop2 for backwards compatibility

The new structure makes atoms independently usable for:
- Temporal workflow activities
- Custom orchestration logic
- Fine-grained execution control
